requirements_pip.txt
pyyaml
librosa
tqdm
torch>=1.3.1
transformers==1.16.4
matplotlib
scikit-image
